Remove unreachable batch sensor endpoints and merge sensor_streaming into sensors
