+++ wrappers/rest-api/app/api/router.py
@@ -2,7 +2,7 @@
 # Copyright(c) 2026 RealSense, Inc. All Rights Reserved.
 
 from fastapi import APIRouter
-from app.api.endpoints import devices, sensors, options, streams, webrtc, point_cloud, firmware, sensor_streaming, hwm, system
+from app.api.endpoints import devices, sensors, options, streams, webrtc, point_cloud, firmware, hwm, system
 
 
 def _get_sdk_version() -> str:
@@ -45,10 +45,3 @@ api_router.include_router(streams.router, prefix="/devices/{device_id}/stream", 
 api_router.include_router(point_cloud.router, prefix="/devices/{device_id}/point_cloud", tags=["point_cloud"])
 api_router.include_router(webrtc.router, prefix="/webrtc", tags=["webrtc"])
 api_router.include_router(system.router, prefix="/system", tags=["system"])
-
-# Per-sensor streaming control (sensor API - independent sensor start/stop)
-api_router.include_router(
-    sensor_streaming.router, 
-    prefix="/devices/{device_id}/sensors", 
-    tags=["sensor-streaming"]
-)
