+++ wrappers/rest-api/tools/react-viewer/src/api/client.ts
@@ -11,9 +11,6 @@ import type {
   ICECandidate,
   SensorStreamConfig,
   SensorStreamStatus,
-  BatchSensorStartRequest,
-  BatchSensorStopRequest,
-  BatchSensorStatus,
 } from './types'
 
 // Detect if running in Tauri desktop app
@@ -234,35 +231,6 @@ class ApiClient {
   async getSensorStatus(deviceId: string, sensorId: string): Promise<SensorStreamStatus> {
     const response = await this.client.get<SensorStreamStatus>(
       `/devices/${deviceId}/sensors/${sensorId}/status`
-    )
-    return response.data
-  }
-
-  async batchStartSensors(
-    deviceId: string,
-    request: BatchSensorStartRequest
-  ): Promise<BatchSensorStatus> {
-    const response = await this.client.post<BatchSensorStatus>(
-      `/devices/${deviceId}/sensors/batch/start`,
-      request
-    )
-    return response.data
-  }
-
-  async batchStopSensors(
-    deviceId: string,
-    request?: BatchSensorStopRequest
-  ): Promise<BatchSensorStatus> {
-    const response = await this.client.post<BatchSensorStatus>(
-      `/devices/${deviceId}/sensors/batch/stop`,
-      request || {}
-    )
-    return response.data
-  }
-
-  async getBatchSensorStatus(deviceId: string): Promise<BatchSensorStatus> {
-    const response = await this.client.get<BatchSensorStatus>(
-      `/devices/${deviceId}/sensors/batch/status`
     )
     return response.data
   }
